heavy refactor
